pull imported content source link helper out of vuex
